SO pipeline with gene and 3'UTR TPM filter and multiple test correction

+++ Riboseq/Riboseq_validation/genomic/resample_random/expressed_genes_multiple_testing_approach_November_2025/genomic_single_samples_resample_random_tpm_rpkm_filter_q_value.sh
@@ -16,7 +16,7 @@ unique_region_dir_ri="/home/ckalk/tools/SplitORF_pipeline/Output/run_30.09.2025-
 ensembl_gtf="/projects/splitorfs/work/reference_files/Homo_sapiens.GRCh38.110.chr.gtf"
 genome_fasta="/projects/splitorfs/work/reference_files/Homo_sapiens.GRCh38.dna.primary_assembly_110.fa"
 # The following are the paths to the riboseq reads
-input_data=(/projects/splitorfs/work/Riboseq/data/fastp/fastp_single_samples/ERR*.fastq)
+input_data=(/projects/splitorfs/work/Riboseq/data/fastp/fastp_single_samples/*.fastq)
 # file with the 3'UTR background regiosn
 three_primes="/projects/splitorfs/work/Riboseq/data/region_input/genomic/3_primes_genomic_merged_numbered.bed"
 
@@ -49,8 +49,8 @@ echo "STAR index genome"
 if [ ! -d  "$output_star"/index ]; then
   source /home/ckalk/scripts/SplitORFs/Riboseq/Riboseq_validation/genomic/resample_random/STAR_Align_genomic_23_09_25.sh \
   -i 50 "$output_star"/index \
-  $genome_fasta \
-  $ensembl_gtf
+  "$genome_fasta" \
+  "$ensembl_gtf"
 fi
 
 echo "Starting alignment against genome"
@@ -60,21 +60,21 @@ echo "Starting alignment against genome"
 for i in "${input_data[@]}"; do
     sample_name=$(basename "$i" _fastp.fastq)
     
-    # if [ ! -e  "$output_star/NMD_genome/${sample_name}/${sample_name}_NMD_chrom_sort.bed" ]; then
+    if [ ! -e  "$output_star/NMD_genome/${sample_name}/${sample_name}_NMD_chrom_sort.bed" ]; then
       echo $i
       mkdir "$output_star"/NMD_genome/${sample_name}
 
       /home/ckalk/scripts/SplitORFs/Riboseq/Riboseq_validation/genomic/resample_random/STAR_Align_genomic_23_09_25.sh \
       -a 16 "$output_star"/index $i \
       "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD \
-      EndToEnd
+      EndToEnd \
+      "$genome_fasta"
 
       rm "$output_star"/NMD_genome/${sample_name}/${sample_name}*Aligned.sortedByCoord.out.bam
       rm "$output_star"/NMD_genome/${sample_name}/${sample_name}*_filtered.bam
-      rm "$output_star"/NMD_genome/${sample_name}/*${sample_name}.bed
 
       echo "===================       Sample $sample_name mapped"
-    # fi
+    fi
 
 done
 
@@ -93,37 +93,35 @@ for folder in "$output_star"/NMD_genome/*/; do
     if [[ -e "${bams[0]}" ]]; then
         i="${bams[0]}"        
         
-        echo "here comes i"
-        echo $i
 
         sample_name="$(basename "$i" _NMD_sorted.bam)"
 
-        # if [ ! -e  "$output_star"/NMD_genome/${sample_name}/Unique_DNA_Regions_genomic_${sample_name}.bed ]; then
-            # if [ ! -e  "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv ]; then
-            #   htseq-count -f bam --secondary-alignments ignore \
-            #     -c  "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
-            #     --supplementary-alignments ignore $i $ensembl_gtf
-            # fi
+        if [ ! -e  "$output_star"/NMD_genome/${sample_name}/Unique_DNA_Regions_genomic_${sample_name}.bed ]; then
+            if [ ! -e  "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv ]; then
+              htseq-count -f bam --secondary-alignments ignore \
+                -c  "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
+                --supplementary-alignments ignore $i $ensembl_gtf
+            fi
 
-            # python filter_bed_file_for_expressed_genes_rnanrom.py \
-            #   $three_primes \
-            #   "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
-            #   $ensembl_gtf \
-            #   2
+            python filter_bed_file_for_expressed_genes_rnanrom.py \
+              $three_primes \
+              "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
+              $ensembl_gtf \
+              5
 
-            # python filter_bed_file_for_expressed_genes_rnanrom.py \
-            #   $unique_region_dir_nmd/Unique_DNA_Regions_genomic.bed \
-            #   "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
-            #   $ensembl_gtf \
-            #   2
+            python filter_bed_file_for_expressed_genes_rnanrom.py \
+              $unique_region_dir_nmd/Unique_DNA_Regions_genomic.bed \
+              "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
+              $ensembl_gtf \
+              5
 
-            # bedtools coverage -s -F 0.33 -a "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_merged_numbered_${sample_name}.bed \
-            # -b $i > "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_merged_numbered_${sample_name}_coverage.tsv
+            bedtools coverage -s -F 0.33 -a "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_merged_numbered_${sample_name}.bed \
+            -b $i > "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_merged_numbered_${sample_name}_coverage.tsv
 
             python filter_tpm_for_3_prime_utrs.py \
             "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_merged_numbered_${sample_name}_coverage.tsv \
             ${sample_name} \
-            100
+            25
 
 
             echo $sample_name
@@ -136,110 +134,138 @@ for folder in "$output_star"/NMD_genome/*/; do
                 ${genome_fasta}
 
             echo "===================       Sample $sample_name intersected"
-        # fi
+        fi
     fi
 
     
 
 done
 
-# # the preprocessing and intersection for Vlado's protocol Ribo-seq data
-# # run this once in the first run, after that there are the chrom_sort.bed files
-# # already in the respective directory, and they are run in the second loop
-# for i in $UMI_dedup_outdir/*_filtered.bam; do
-#     sample_name=$(basename "$i" _dedup_filtered.bam)
+# the preprocessing and intersection for Vlado's protocol Ribo-seq data
+# run this once in the first run, after that there are the chrom_sort.bed files
+# already in the respective directory, and they are run in the second loop
+for i in $UMI_dedup_outdir/*_filtered.bam; do
+    sample_name=$(basename "$i" _dedup_filtered.bam)
     
-#     echo $sample_name
+    echo $sample_name
 
-#     if [[ "$sample_name" == uf_mueller* ]];then
-#         sample_name="${sample_name:30}"
-#     fi
+    if [[ "$sample_name" == uf_mueller* ]];then
+        sample_name="${sample_name:30}"
+    fi
 
-#     if [ ! -e  "$output_star"/NMD_genome/${sample_name}/Unique_DNA_Regions_genomic_${sample_name}.bed ]; then
+    if [ ! -e  "$output_star"/NMD_genome/${sample_name}/Unique_DNA_Regions_genomic_${sample_name}.bed ]; then
 
-#         mkdir "$output_star"/NMD_genome/${sample_name}
+        if [ ! -d "$output_star"/NMD_genome/${sample_name} ];then
+            mkdir "$output_star"/NMD_genome/${sample_name}
+        fi
 
-#         htseq-count -f bam --secondary-alignments ignore \
-#           -c  "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
-#           --supplementary-alignments ignore $i $ensembl_gtf
+        if [ ! -e  "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv ]; then
+          htseq-count -f bam --secondary-alignments ignore \
+            -c  "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
+            --supplementary-alignments ignore $i $ensembl_gtf
+        fi
 
-#         python filter_bed_file_for_expressed_genes_rnanrom.py \
-#           $three_primes \
-#           "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
-#           $ensembl_gtf \
-#           2
+            python filter_bed_file_for_expressed_genes_rnanrom.py \
+              $three_primes \
+              "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
+              $ensembl_gtf \
+              5
 
-#         python filter_bed_file_for_expressed_genes_rnanrom.py \
-#           $unique_region_dir_nmd/Unique_DNA_Regions_genomic.bed \
-#           "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
-#               2
+            python filter_bed_file_for_expressed_genes_rnanrom.py \
+              $unique_region_dir_nmd/Unique_DNA_Regions_genomic.bed \
+              "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
+              $ensembl_gtf \
+              5
 
-#         echo $sample_name
-#         ./empirical_intersection_steps_expressed_genes_filter_18_11_25.sh  \
-#             $i \
-#             "$output_star"/NMD_genome/${sample_name}/Unique_DNA_Regions_genomic_${sample_name}.bed \
-#             "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_${sample_name}.bed \
-#             "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD \
-#             "$output_star"/NMD_genome/${sample_name} \
-#             ${genome_fasta}
+        bedtools coverage -s -F 0.33 -a "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_merged_numbered_${sample_name}.bed \
+            -b $i > "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_merged_numbered_${sample_name}_coverage.tsv
 
-#         echo "===================       Sample $sample_name intersected"
-#     fi
-
-# done
+        python filter_tpm_for_3_prime_utrs.py \
+            "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_merged_numbered_${sample_name}_coverage.tsv \
+            ${sample_name} \
+            25
 
 
-# # The preprocessing and intersection for the UPF1 deletion Ribo-seq data
-# for i in $umi_dedup_upf10/*_filtered.bam; do
-#     sample_name=$(basename "$i" _dedup_filtered.bam)
+        echo $sample_name
+        ./empirical_intersection_steps_expressed_genes_filter_18_11_25.sh  \
+                $i \
+                "$output_star"/NMD_genome/${sample_name}/Unique_DNA_Regions_genomic_${sample_name}.bed \
+                "$output_star"/NMD_genome/${sample_name}/3_primes_tpm_filtered_${sample_name}.bed \
+                "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD \
+                "$output_star"/NMD_genome/${sample_name} \
+                ${genome_fasta}
+
+        echo "===================       Sample $sample_name intersected"
+    fi
+
+done
+
+
+# The preprocessing and intersection for the UPF1 deletion Ribo-seq data
+for i in $umi_dedup_upf10/*_filtered.bam; do
+    sample_name=$(basename "$i" _dedup_filtered.bam)
     
-#     echo $sample_name
+    echo $sample_name
 
-#     if [[ "$sample_name" == uf_mueller* ]];then
-#         sample_name="${sample_name:30}"
-#     fi
+    if [[ "$sample_name" == uf_mueller* ]];then
+        sample_name="${sample_name:30}"
+    fi
 
-# #     if [ ! -e  "$output_star"/NMD_genome/${sample_name}/Unique_DNA_Regions_genomic_${sample_name}.bed ]; then
-# #         mkdir "$output_star"/NMD_genome/${sample_name}
+    if [ ! -e  "$output_star"/NMD_genome/${sample_name}/Unique_DNA_Regions_genomic_${sample_name}.bed ]; then
+        if [ ! -d "$output_star"/NMD_genome/${sample_name} ];then
+            mkdir "$output_star"/NMD_genome/${sample_name}
+        fi
+        
+        if [ ! -e  "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv ]; then
+            htseq-count -f bam --secondary-alignments ignore \
+              -c  "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
+              --supplementary-alignments ignore $i $ensembl_gtf
+        fi
 
-# #         htseq-count -f bam --secondary-alignments ignore \
-# #           -c  "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
-# #           --supplementary-alignments ignore $i $ensembl_gtf
+            python filter_bed_file_for_expressed_genes_rnanrom.py \
+              $three_primes \
+              "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
+              $ensembl_gtf \
+              5
 
-#         python filter_bed_file_for_expressed_genes_rnanrom.py \
-#           $three_primes \
-#           "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
-#           $ensembl_gtf \
-#           2
+            python filter_bed_file_for_expressed_genes_rnanrom.py \
+              $unique_region_dir_nmd/Unique_DNA_Regions_genomic.bed \
+              "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
+              $ensembl_gtf \
+              5
 
-#         python filter_bed_file_for_expressed_genes_rnanrom.py \
-#           $unique_region_dir_nmd/Unique_DNA_Regions_genomic.bed \
-#           "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD_htseq_counts.tsv \
-#               2
+        bedtools coverage -s -F 0.33 -a "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_merged_numbered_${sample_name}.bed \
+            -b $i > "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_merged_numbered_${sample_name}_coverage.tsv
 
-# #         echo $sample_name
-# #         ./empirical_intersection_steps_expressed_genes_filter_18_11_25.sh  \
-# #             $i \
-# #             "$output_star"/NMD_genome/${sample_name}/Unique_DNA_Regions_genomic_${sample_name}.bed \
-# #             "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_${sample_name}.bed \
-# #             "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD \
-# #             "$output_star"/NMD_genome/${sample_name} \
-# #             ${genome_fasta}
-
-# #         echo "===================       Sample $sample_name intersected"
-# #     fi
-
-# done
-
+        python filter_tpm_for_3_prime_utrs.py \
+            "$output_star"/NMD_genome/${sample_name}/3_primes_genomic_merged_numbered_${sample_name}_coverage.tsv \
+            ${sample_name} \
+            25
 
 
+        echo $sample_name
+        ./empirical_intersection_steps_expressed_genes_filter_18_11_25.sh  \
+                $i \
+                "$output_star"/NMD_genome/${sample_name}/Unique_DNA_Regions_genomic_${sample_name}.bed \
+                "$output_star"/NMD_genome/${sample_name}/3_primes_tpm_filtered_${sample_name}.bed \
+                "$output_star"/NMD_genome/${sample_name}/${sample_name}_NMD \
+                "$output_star"/NMD_genome/${sample_name} \
+                ${genome_fasta}
 
-# export LD_LIBRARY_PATH=/opt/intel/oneapi/mkl/2022.0.2/lib/intel64:$LD_LIBRARY_PATH
-# export MKL_ENABLE_INSTRUCTIONS=SSE4_2
+        echo "===================       Sample $sample_name intersected"
+    fi
 
-# Rscript -e 'if (!requireNamespace("rmarkdown", quietly = TRUE)) install.packages("rmarkdown", repos="http://cran.us.r-project.org")'
+done
 
-# R -e 'library(rmarkdown); rmarkdown::render(input = "RiboSeqReportGenomic_iteration_update_expression_filter_multiple_test_correction.Rmd", output_file = "/projects/splitorfs/work/Riboseq/Output/Riboseq_genomic_single_samples/resample_q10_expression_filter/Riboseq_report_NMD_expression_filter_multiple_test_correction.pdf", params=list(args = c("/projects/splitorfs/work/Riboseq/Output/Riboseq_genomic_single_samples/resample_q10_expression_filter/NMD_genome", "/home/ckalk/tools/SplitORF_pipeline/Output/run_30.09.2025-11.30.56_NMD_transcripts_correct_TSL_ref", "NMD")))'
+
+
+
+export LD_LIBRARY_PATH=/opt/intel/oneapi/mkl/2022.0.2/lib/intel64:$LD_LIBRARY_PATH
+export MKL_ENABLE_INSTRUCTIONS=SSE4_2
+
+Rscript -e 'if (!requireNamespace("rmarkdown", quietly = TRUE)) install.packages("rmarkdown", repos="http://cran.us.r-project.org")'
+
+R -e 'library(rmarkdown); rmarkdown::render(input = "RiboSeqReportGenomic_iteration_update_expression_filter_multiple_test_correction.Rmd", output_file = "/projects/splitorfs/work/Riboseq/Output/Riboseq_genomic_single_samples/resample_q10_expression_filter/Riboseq_report_NMD_tpm_filter_multiple_test_correction_08_12_25.pdf", params=list(args = c("/projects/splitorfs/work/Riboseq/Output/Riboseq_genomic_single_samples/resample_q10_expression_filter/NMD_genome", "/home/ckalk/tools/SplitORF_pipeline/Output/run_30.09.2025-11.30.56_NMD_transcripts_correct_TSL_ref", "NMD")))'
